DETR_Vehicle_counting/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
easydict 
thop
google-generativeai>=0.3.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0 